fix: computer-metrics live-response package
